[SPARK-50255][PYTHON] Avoid unnecessary casting in `compute_hist`
### What changes were proposed in this pull request?
Avoid unnecessary casting in `compute_hist`

### Why are the changes needed?
the `__bucket` should be integer by its nature, it was double just because of the output type of `Bucketizer` in MLlib (almost all ML implementations returns double in transformation).
After reimplementing it with Spark SQL, it no longer needs to be float.

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
CI

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #48785 from zhengruifeng/plt_hist_cast.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>