[SPARK-50238][PYTHON] Add Variant Support in PySpark UDFs/UDTFs/UDAFs
### What changes were proposed in this pull request?

This PR adds support for the Variant type in PySpark UDFs/UDTFs/UDAFs. Support is added in both modes - arrow and pickle - and support is also added in pandas UDFs.

### Why are the changes needed?

After this change, users will be able to use the new Variant data type with UDFs, which is currently prohibited.

### Does this PR introduce _any_ user-facing change?

Yes, users should now be able to use Variants with Python UDFs.

### How was this patch tested?

Unit tests in all scenarios - arrow, pickle and pandas

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48770 from harshmotw-db/harsh-motwani_data/variant_udf_3.

Authored-by: Harsh Motwani <harsh.motwani@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>