[SPARK-49611][SQL][FOLLOW-UP] Make collations TVF consistent and return null on no result for country and language
### What changes were proposed in this pull request?
It was noticed that we return null for country and language for collations TVF when collation is UTF8_*, but when information is missing in ICU we return empty string.

### Why are the changes needed?
Making behaviour consistent.

### Does this PR introduce _any_ user-facing change?
No, this is all in Spark 4.0, so addition of this TVF was not released yet.

### How was this patch tested?
Existing test.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48835 from mihailom-db/fix-collations-table.

Authored-by: Mihailo Milosevic <mihailo.milosevic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>