[SPARK-50092][SQL] Fix PostgreSQL connector behaviour for multidimensional arrays
### What changes were proposed in this pull request?
There is a bug introduced in this PR https://github.com/apache/spark/pull/46006. This PR fixed the behaviour for PostgreSQL connector for multidimensional arrays since we have mapped all arrays to 1D arrays.

This PR has introduced a bug for one case. Following scenario is broken:

- User has a table t1 on Postgres and does CTAS command to create table t2 with same data.
PR 46006 is resolving the dimensionality of column by reading the metadata from pg_attribute table and attndims column.
- This query returns correct dimensionality for table t1, but for table t2 that is created via CTAS it returns 0 always. This leads to all of the arrays being mapped to 0-D array which is the type itself (for example int). This is a bug on Postgres side.
- As a solution, we can query array_ndims function on given column that will return the dimension of the column. It works for CTAS-like-created tables too. We can get the result of this function on first row of table.

This change is doing additional query to PG table to find the dimension of array column instead of querying metadata table as before. It might be more expensive but we are sending LIMIT 1 in query.

Also, there is one caveat. In PG, there is no dimensionality of array as all the arrays are 1D arrays (https://www.postgresql.org/docs/current/arrays.html#ARRAYS-DECLARATION). Therefore, if there is table with 2D array column, it is totally fine from PG side to insert 1D or 3D array in this column. This makes the read path on Spark problematic since we will get the dimension of array, for example 1 if the first record is 1D array, and then we will try to read 3D array later on which will fail. Vice versa also, getting dimension 3 and reading 1D array later on. The change that I propose is fine with this scenario since it already doesn't work in Spark. What my change implies is that user can get different error message depending on the dimensionality of first record in table (namely, for one table they can get the error message that the expected type is ARRAY<ARRAY<INT>> and for the other that it is ARRAY\<INT\>.

### Why are the changes needed?
Bug fix.

### Does this PR introduce _any_ user-facing change?
No. It just fixes one case that doesn't work currently.

### How was this patch tested?
New tests.

### Was this patch authored or co-authored using generative AI tooling?

Closes #48625 from PetarVasiljevic-DB/fix_postgres_multidimensional_arrays.

Authored-by: Petar Vasiljevic <petar.vasiljevic@databricks.com>
Signed-off-by: Kent Yao <yao@apache.org>