[SPARK-50216][SQL][TESTS] Update `CollationBenchmark` to invoke `collationNameToId` outside of cases
### What changes were proposed in this pull request?
In this PR, UTF8_BINARY performance regression is addressed, that was first identified here https://github.com/apache/spark/pull/48721. The regression is traced back to this PR https://github.com/apache/spark/pull/48222 when it first occurred, however this isn't the actual source of performance degradation.

### Why are the changes needed?
The PR https://github.com/apache/spark/pull/48222 caused the regression because it changed the `collationNameToId` function and made it slightly slower by removing a short-circuit for fetching the UTF8_BINARY collation. However this function should be called fixed amount of times for each query and from the benchmark framework at most once - this was not the case and it was the largest contributor to performance regression.

This PR addresses the benchmarking framework to not call this function at each expression, but once per the test case.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing testing surface, benchmarks.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #48804 from stevomitric/stevomitric/fix-utf8_binary-regression.

Authored-by: Stevo Mitric <stevo.mitric@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>