[SPARK-50313][SQL][TESTS][FOLLOWUP] Restore some tests in *SQLQueryTestSuite for ANSI mode
### What changes were proposed in this pull request?
This PR makes a follow-up for SPARK-50313, to restore some tests in *SQLQueryTestSuite for ANSI mode. These tests are originally defined in ansi/ folder, in other words, they are not `--IMPORT`-ed, so we need to move them to the default folder where now defaults to ANSI after SPARK-50313.

### Why are the changes needed?
Improve ANSI test coverage.

### Does this PR introduce _any_ user-facing change?
NO

### How was this patch tested?

*SQLQueryTestSuite

### Was this patch authored or co-authored using generative AI tooling?
NO

Closes #48872 from yaooqinn/SPARK-50313-F.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>