[SPARK-50381][CORE] Support `spark.master.rest.maxThreads`
### What changes were proposed in this pull request?

This PR aims to support `spark.master.rest.maxThreads`.

### Why are the changes needed?

To provide users a way to control the number of maximum threads of REST API. Previously, Apache Spark uses a default constructor whose value is fixed to `200` always.

https://github.com/apache/spark/blob/2e1c3dc8004b4f003cde8dfae6857f5bef4bb170/core/src/main/scala/org/apache/spark/deploy/rest/RestSubmissionServer.scala#L94

https://github.com/jetty/jetty.project/blob/5dfc59a691b748796f922208956bd1f2794bcd16/jetty-util/src/main/java/org/eclipse/jetty/util/thread/QueuedThreadPool.java#L118-L121

### Does this PR introduce _any_ user-facing change?

No, the default value of new configuration is identical with the previously-used Jetty's default value.

### How was this patch tested?

Pass the CIs with a newly added test case.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48921 from dongjoon-hyun/SPARK-50381.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>