[SPARK-50287][SQL] Merge options of table and relation when creating WriteBuilder in FileTable
### What changes were proposed in this pull request?

Merge `options` of table and relation when creating WriteBuilder in FileTable.

### Why are the changes needed?

Similar to SPARK-49519 which fixes the read path.

### Does this PR introduce _any_ user-facing change?

FileTable's options are accounted on the V2 write path now, but given the built-in file formats use V1 by default, it has no real effect.

### How was this patch tested?

UT is updated to cover the case.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48821 from pan3793/SPARK-50287.

Authored-by: Cheng Pan <chengpan@apache.org>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>