[SPARK-48344][SQL] Add SQL Scripting Execution Framework
### What changes were proposed in this pull request?
This PR is second in series of refactoring Initial refactoring of SQL Scripting to prepare it for addition of **Execution Framework**:
- Introducing `SqlScriptExecution`, new iterator that collects and returns only results.
- Enabling execution of SQL Scripting using `sql()` API.
- Enabling named parameters to be used with SQL Scripting.

### Why are the changes needed?
This changes are needed to enable execution of SQL Scripts.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
New `SqlScriptingExecutionSuite` to test behavior of newly added component.
New `SqlScriptingE2eSuite` to test end to end behavior using `sql()` API.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48950 from miland-db/milan-dankovic_data/refactor-execution-2.

Authored-by: Milan Dankovic <milan.dankovic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>