[SPARK-50437][SS] Reduce overhead of creating deserializers in TransformWithStateExec
### What changes were proposed in this pull request?

An optimization that creates deserializers for keys and values at the beginning of a batch.

### Why are the changes needed?
Currently, deserializers for key and values are created every time `handleInputRows` gets called.   The function
`ObjectOperator.deserializeRowToObject` is called twice do generate the deserializer for the key and value.  The function will do the following things
1. Generate java code for the deserializer based on input expression and grouping attributes
2. Compile that java code to java class

The problem is that step 1 will occur every time the function is called.  The compiling the code, i.e. step 2, will only happen once since the generated class will be cached and can be reused.

Step 1 can incur a heavy penalty in situations where handleInputRows will be called many times.  For example, if there are many distinct keys.  Note that handleInputRows is called once per distinct key.

Here is a flamegraph:

![Uploading Screenshot 2024-11-26 at 11.55.12 PM.png…]()

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?

Existing tests should suffice

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #48985 from jerrypeng/SPARK-50437.

Authored-by: Jerry Peng <jerry.peng@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>