[SPARK-50194][SS][PYTHON] Integration of New Timer API and Initial State API with Timer
### What changes were proposed in this pull request?

As Scala side, we modify the timer API with a separate `handleExpiredTimer` function inside `StatefulProcessor`, this PR make a change to the timer API to couple with API on Scala side. Also adds a timer parameter to pass into `handleInitialState` function to support use cases for registering timers in the first batch for initial state rows.

### Why are the changes needed?

This change is to couple with Scala side of APIs: https://github.com/apache/spark/pull/48553

### Does this PR introduce _any_ user-facing change?

Yes.
We add a new user defined function to explicitly handle expired timeres:
```
def handleExpiredTimer(
        self, key: Any, timer_values: TimerValues, expired_timer_info: ExpiredTimerInfo
```
We also add a new timer parameter to enable users to register timers for keys exist in the initial state:
```
def handleInitialState(
        self,
        key: Any,
        initialState: "PandasDataFrameLike",
        timer_values: TimerValues) -> None
```

### How was this patch tested?

Add a new test in `test_pandas_transform_with_state`

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #48838 from jingz-db/python-new-timer.

Lead-authored-by: jingz-db <jing.zhan@databricks.com>
Co-authored-by: Jing Zhan <135738831+jingz-db@users.noreply.github.com>
Co-authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>