[SPARK-50430][CORE][FOLLOW-UP] Keep the logic of manual putting key and values in Properties
This PR proposes to actually more conservatively preserve the original code of creating new properties instead of cloning.

Previous codes only copied the key and values but `clone` actually copies more fields in `Properties`. `cloneProperties` is being used in Spark Core, and all other components so I propose to keep the logic as is.

This is more a fix of a potential bug.

No, it is difficult to add a test.

No.

Closes #49036 from HyukjinKwon/SPARK-50430-followup.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit 4abaab3ffeba5a3d39216e7224928bb82b254e22)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>