[SPARK-50473][SQL] Simplify classic Column handling
### What changes were proposed in this pull request?
We added a couple of helper functions that make it easier to work Columns in the Classic API. This covers functionality that creates a Column from an Expression, and creating (named) Expressions from a Column. There are currently multiple ways of doing the same thing and this is confusing. This PR attempts to simplify this a bit, by making the following changes:

- `ExpressionUtils` is moved to the background. `ClassicConversions`/`ColumnConversion` are now predominantly used. The benefit of this is that most code now looks like pre-Spark 4 code.
- `ExpressionUtils.expression(..)` and `ExpressionUtils.column(..)` are not implicit anymore. This was confusing.
- `testImplicits` now supports both Expression -> Column and Column -> Expression conversions.

### Why are the changes needed?
Easier to understand code.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #49038 from hvanhovell/SPARK-50473.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>