[SPARK-49695][SQL] Postgres fix xor push-down
### What changes were proposed in this pull request?
This PR fixes the pushdown of ^ operator (XOR operator) for Postgres. Those two databases use this as exponent, rather then bitwise xor.

Fix is consisted of overriding the SQLExpressionBuilder to replace the '^' character with '#'.
### Why are the changes needed?
Result is incorrect.

### Does this PR introduce _any_ user-facing change?
Yes. The user will now have a proper translation of the ^ operator.

### How was this patch tested?

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48144 from andrej-db/SPARK-49695-PostgresXOR.

Lead-authored-by: Andrej Gobeljić <andrej.gobeljic@databricks.com>
Co-authored-by: andrej-db <andrej.gobeljic@databricks.com>
Co-authored-by: andrej-gobeljic_data <andrej.gobeljic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>